+++ src/irina/youtube.rs
@@ -5,6 +5,8 @@
 use crate::feed::Entry;
 use crate::youtube::{Feed, YoutubeClient};
 
+use chrono::{DateTime, Utc};
+
 // <https://www.youtube.com/feeds/videos.xml?channel_id=UCK3cMi97Kf_WENLvRFdztoQ>
 const CHANNEL_ID: &str = "UCK3cMi97Kf_WENLvRFdztoQ";
 
@@ -16,8 +18,24 @@ impl Youtube {
         if let Some(video) = Self::get_latest_videos().await?.entries().next() {
             Ok(video.clone())
         } else {
-            anyhow::bail!("Ciao sono Irina. Non ho trovato nessun video per te")
+            anyhow::bail!("Ciao sono Irina. Non ho nessun video da mostrarti.")
         }
+    }
+
+    /// Get oldest unseen video from youtube
+    pub async fn get_oldest_unseen_video(
+        last_video_pubdate: DateTime<Utc>,
+    ) -> anyhow::Result<Option<Entry>> {
+        let feed = Self::get_latest_videos().await?;
+        // sort by date
+        let mut entries: Vec<Entry> = feed.entries().cloned().collect();
+        entries.sort_by_key(|x| x.date);
+        for entry in entries.into_iter() {
+            if entry.date > Some(last_video_pubdate) {
+                return Ok(Some(entry));
+            }
+        }
+        Ok(None)
     }
 
     /// Get latest videos from spazio grigio
@@ -25,7 +43,7 @@ impl Youtube {
         let client = YoutubeClient::new(CHANNEL_ID);
         client.fetch().await.map_err(|e| {
             anyhow::anyhow!(
-                "Ciao sono Irina e non riesco a darti i miei ultimi video: {}",
+                "Ciao sono Irina. Non riesco ad ottenere gli ultimi video di Spazio Grigio: {}",
                 e
             )
         })