+++ CHANGELOG.md
@@ -1,12 +1,21 @@
 # Changelog
 
 - [Changelog](#changelog)
+  - [0.3.0](#030)
   - [0.2.1](#021)
   - [0.2.0](#020)
   - [0.1.1](#011)
   - [0.1.0](#010)
 
 ---
+
+## 0.3.0
+
+Released on 05/09/2022
+
+- Revert to picuki
+- Fetch for latest unseen instagram post, instead of latest in automatizer.
+- Fetch for latest unseen youtube video, instead of latest in automatizer.
 
 ## 0.2.1